% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repacked_funs.R
\name{ExtractPIPs_parallel}
\alias{ExtractPIPs_parallel}
\title{Posterior inclusion probabilities by chain}
\usage{
ExtractPIPs_parallel(x, ...)
}
\arguments{
\item{x}{bkmrfit.list object from \code{\link[bkmrhat]{kmbayes_parallel}}}

\item{...}{arguments to \code{\link[bkmr]{CalcPIPs}}}
}
\value{
data.frame with all chains together
}
\description{
Posterior inclusion probabilities by chain
}
