% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repacked_funs.R
\name{OverallRiskSummaries_parallel}
\alias{OverallRiskSummaries_parallel}
\title{Overall summary by chain}
\usage{
OverallRiskSummaries_parallel(x, ...)
}
\arguments{
\item{x}{bkmrfit.list object from \code{\link[bkmrhat]{kmbayes_parallel}}}

\item{...}{arguments to \code{\link[bkmr]{OverallRiskSummaries}}}
}
\value{
data.frame with all chains together
}
\description{
Overall summary by chain
}
