% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repacked_funs.R
\name{PredictorResponseBivar_parallel}
\alias{PredictorResponseBivar_parallel}
\title{Bivariate predictor response by chain}
\usage{
PredictorResponseBivar_parallel(x, ...)
}
\arguments{
\item{x}{bkmrfit.list object from \code{\link[bkmrhat]{kmbayes_parallel}}}

\item{...}{arguments to \code{\link[bkmr]{PredictorResponseBivar}}}
}
\value{
data.frame with all chains together
}
\description{
Bivariate predictor response by chain
}
