% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repacked_funs.R
\name{PredictorResponseUnivar_parallel}
\alias{PredictorResponseUnivar_parallel}
\title{Univariate predictor response summary by chain}
\usage{
PredictorResponseUnivar_parallel(x, ...)
}
\arguments{
\item{x}{bkmrfit.list object from \code{\link[bkmrhat]{kmbayes_parallel}}}

\item{...}{arguments to \code{\link[bkmr]{PredictorResponseUnivar}}}
}
\value{
data.frame with all chains together
}
\description{
Univariate predictor response summary by chain
}
