% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repacked_funs.R
\name{SamplePred_parallel}
\alias{SamplePred_parallel}
\title{Posterior samples of E(Y|h(Z),X,beta) by chain}
\usage{
SamplePred_parallel(x, ...)
}
\arguments{
\item{x}{bkmrfit.list object from \code{\link[bkmrhat]{kmbayes_parallel}}}

\item{...}{arguments to \code{\link[bkmr]{CalcPIPs}}}
}
\value{
data.frame with all chains together
}
\description{
Posterior samples of E(Y|h(Z),X,beta) by chain
}
