% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repacked_funs.R
\name{SingVarRiskSummaries_parallel}
\alias{SingVarRiskSummaries_parallel}
\title{Single variable summary by chain}
\usage{
SingVarRiskSummaries_parallel(x, ...)
}
\arguments{
\item{x}{bkmrfit.list object from \code{\link[bkmrhat]{kmbayes_parallel}}}

\item{...}{arguments to \code{\link[bkmr]{SingVarRiskSummaries}}}
}
\value{
data.frame with all chains together
}
\description{
Single variable summary by chain
}
